requirements.txt
# Django
# --------------------------------
Django==4.0.6
djangorestframework==3.13.1
drf-spectacular==0.22.1